+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Stream chunks are written to stdout incrementally instead of buffering until drain completes
+
 ## [0.18.10] - 2026-04-09
 
 ### Added